==> onsides_intl/onsides_uk/notebooks/3-1.term_extraction_ade_serious_ade.ipynb <==
import numpy as np
import pandas as pd
import requests
from tqdm import tqdm
import ast, re
from time import sleep
from bs4 import BeautifulSoup
data_folder = '/content/drive/MyDrive/pop_pharmacogenomics/onsides_intl/uk_drug_label/'
external_data_folder = '/content/drive/MyDrive/pop_pharmacogenomics/onsides_intl/external_data/'
api = 'bbf40496-e20e-4f79-9636-b63d4a3deea7'
import warnings
warnings.filterwarnings('ignore')

### Extract terms from text data here

text_df = pd.read_csv(data_folder+'data/drug_wp_data_raw.csv')
text_df['section_title'] = text_df['drug_text'].apply(lambda x: BeautifulSoup(x, 'html.parser').find('summary').text)
text_df['section_content'] = text_df.apply(lambda x: BeautifulSoup(x.drug_text, 'html.parser').text.replace(x.section_title, '').strip().lower(),
                                           axis = 1)
text_df.head(1)

unique_text = text_df[['section_content']].drop_duplicates()
print(unique_text.shape[0])

#read in the meddra map file
f = '/content/drive/MyDrive/pop_pharmacogenomics/onsides_intl/'
meddra_df = pd.read_csv(f+'external_data/umls_meddra_en.csv')
meddra_df['STR'] = meddra_df.STR.apply(lambda x: x.lower())
meddra_df['len'] = meddra_df.STR.apply(lambda x: len(x))
print(meddra_df.shape[0])
meddra_df = meddra_df[(meddra_df['len'] >= 5)]
meddra_df = meddra_df[(meddra_df.TTY.isin(['PT', 'LLT']))]
print(meddra_df.shape[0])
meddra_dict = dict(zip(meddra_df.STR, meddra_df.SDUI))
meddra_names = meddra_df.STR.tolist()
meddra_df.head(1)

app_data = []
for ade_text in tqdm(text_df.section_content.tolist()):
  if str(ade_text) != 'nan':
    ade_text = ade_text.lower()
    meddra_found_terms = []
    #iterate through list of meddra concepts, if found - add to list
    for concept_name in meddra_names:
      if ade_text.find(concept_name) == -1:
        continue
      else:
        i = ade_text.index(concept_name)
        meddra_found_terms.append([i, concept_name, meddra_dict[concept_name]])
    app_data.append([ade_text, meddra_found_terms])

app_data_df = pd.DataFrame(app_data, columns = ['string', 'list'])
app_data_df = app_data_df.explode('list')
app_data_df['index'] = app_data_df['list'].apply(lambda x: x[0] if str(x) != 'nan' else None)
app_data_df['found_term'] = app_data_df['list'].apply(lambda x: x[1] if str(x) != 'nan' else None)
app_data_df['meddra_id'] = app_data_df['list'].apply(lambda x: x[2] if str(x) != 'nan' else None)
app_data_df = app_data_df.drop(['list'], axis = 1)
app_data_df.to_csv(data_folder+'data/drug_wp_text_matched_terms.csv', index=False)
app_data_df.head(1)

### prepare onsides input

text_df = pd.read_csv(data_folder+'data/drug_wp_data_raw.csv')
text_df['section_title'] = text_df['drug_text'].apply(lambda x: BeautifulSoup(x, 'html.parser').find('summary').text)
text_df['section_content'] = text_df.apply(lambda x: BeautifulSoup(x.drug_text, 'html.parser').text.replace(x.section_title, '').strip().lower(),
                                           axis = 1)
text_df.head(1)

app_data_df = pd.read_csv(data_folder+'data/drug_wp_text_matched_terms.csv')
print(app_data_df.shape)
app_data_df = app_data_df.drop_duplicates()
print(app_data_df.shape)
app_data_df.head(1)

building_strings = []
#ade_text_table_dict = dict(zip(text_df.product_id, text_df.section_content))
for i, row in tqdm(app_data_df.iterrows()):
  if str(row['index']) != 'nan':
    term, start_pos, ar_text = row['found_term'], int(row['index']), row['string']
    #default settings
    nwords, prop_before = 125, 0.125
    #pull the full text
    #ar_text = ade_text_table_dict[label_id]

    term_nwords = len(term)
    size_before = max(int((nwords-2*term_nwords)*prop_before), 1)
    size_after = max(int((nwords-2*term_nwords)*(1-prop_before)), 1)

    before_text = ar_text[:start_pos]
    after_text = ar_text[(start_pos+term_nwords):]

    before_parts = before_text.split()[-1*size_before:]
    after_parts = after_text.split()[:size_after]

    li = [term]
    li.extend(before_parts)
    li.append('EVENT')
    li.extend(after_parts)
    example_string = ' '.join(li)
    building_strings.append(example_string)
  else:
    building_strings.append(None)

app_data_df['section_content'] = app_data_df['string']
app_data_df['string'] = building_strings
exact_terms_df = text_df.merge(app_data_df, on = 'section_content', how = 'left')
exact_terms_df = exact_terms_df.drop(columns = ['drug_text', 'section_content', 'section_title'])
exact_terms_df.head()

exact_terms_df.to_csv(data_folder+'data/serious/sentences-rx_method14_nwords125_clinical_bert_application_set_AR_v0120.csv', index=False)

---
### format dataframe for onsides model

#required columns : section, drug, label_id, set_id, spl_version, pt_meddra_id, pt_meddra_term
exact_terms_df = pd.read_csv(data_folder+'data/serious/sentences-rx_method14_nwords125_clinical_bert_application_set_AR_v0120.csv')
exact_terms_df['section'] = 'AR'
exact_terms_df['set_id'] = exact_terms_df['product_id']
exact_terms_df['location'] = exact_terms_df['index']
exact_terms_df = exact_terms_df.rename(columns = {'product_id':'label_id', 'onsides_string':'string'}).drop(columns = ['index'])
print(exact_terms_df.shape)
drug_df = pd.read_csv(data_folder+'data/final_UK/drug_data.csv').rename(columns={'ingredient_name':'drug'})
drug_df['label_id'] = drug_df['product_id'].apply(lambda x: int(x.split('/')[-2]))
drug_df = drug_df[['label_id', 'drug']].drop_duplicates().groupby('label_id')['drug'].apply(set).reset_index()
drug_df['drug'] = drug_df['drug'].apply(lambda x: str(x)[1:-1].replace("'",''))
exact_terms_df = exact_terms_df.merge(drug_df[['label_id', 'drug']].drop_duplicates(), on = 'label_id', how = 'left')
print(exact_terms_df.shape)
folder = '/content/drive/MyDrive/pop_pharmacogenomics/fda/onsides-2.0.0/data/'
llt_pt = pd.read_csv(folder+'meddra_llt_pt_map.txt', delimiter = '|')
llt_pt_id_dict = dict(zip(llt_pt.llt_concept_code, llt_pt.pt_concept_code))
llt_pt_term_dict = dict(zip(llt_pt.llt_concept_code, llt_pt.pt_concept_name))
exact_terms_df['pt_meddra_id'] = exact_terms_df.meddra_id.apply(lambda x: llt_pt_id_dict[x] if x in llt_pt_id_dict.keys() else None)
exact_terms_df['pt_meddra_term'] =  exact_terms_df.meddra_id.apply(lambda x: llt_pt_term_dict[x] if x in llt_pt_term_dict.keys() else None)
print(exact_terms_df.shape)
exact_terms_df.to_csv(data_folder+'data/serious/sentences-rx_method14_nwords125_clinical_bert_application_set_AR_v0120.csv', index=False)
exact_terms_df.head(1)

exact_terms_df = pd.read_csv(data_folder+'data/serious/sentences-rx_method14_nwords125_clinical_bert_application_set_AR_v0120.csv')
print(exact_terms_df.shape)
exact_terms_df = exact_terms_df[exact_terms_df.string.notna()]
exact_terms_df['spl_version'] = None
print(exact_terms_df.shape)
exact_terms_df.to_csv(data_folder+'data/serious/sentences-rx_method14_nwords125_clinical_bert_application_set_AR_v0120.csv', index=False)

### run onsides model

#prep onsides materials
folder = '/content/drive/MyDrive/pop_pharmacogenomics/onsides_intl/onsides'
ar_model = folder + '/bestepoch-bydrug-PMB_14-AR-125-all_222_24_25_2.5e-05_256_32.pth'
bw_model = folder + '/bestepoch-bydrug-PMB_14-AR-125-all_222_24_25_2.5e-05_256_32.pth'
import os
os.chdir('/content/drive/MyDrive/pop_pharmacogenomics/fda/onsides-2.0.0/')
!python3 -m pip install -r requirements.txt

#use onsides model
f = data_folder+'data/serious/sentences-rx_method14_nwords125_clinical_bert_application_set_AR_v0120.csv'
!python3 src/predict.py --model $ar_model --examples $f

#build tables
r = data_folder+'data/serious/bestepoch-bydrug-PMB-sentences-rx_ref14-AR-125-all_222_24_25_2.5e-05_256_32.csv.gz'
!python3 src/create_onsides_datafiles.py --release v2.0.0-AR --results $r --examples $f

res = data_folder+'data/bestepoch-bydrug-PMB-sentences-rx_ref14-AR-125-all_222_24_25_2.5e-05_256_32.csv.gz'
ex = data_folder+'data/sentences-rx_method14_nwords125_clinical_bert_application_set_AR_v0924.csv'
#threshold = 0.463
res = pd.read_csv(res, header=None, names=['Pred0', 'Pred1'])
ex = pd.read_csv(ex)
df = pd.concat([ex, res], axis=1)
print(df.shape[0])
df = df[df.Pred0 > threshold]
print(df.shape[0])
df.head()

df.to_csv(data_folder+'data/ade_text_table_onsides_pred_v0924.csv', index=False)
