==> onsides_intl/onsides_uk/notebooks/3-2.term_extraction_specific_populations.ipynb <==
import numpy as np
import pandas as pd
import requests
from tqdm import tqdm
import ast, re
from time import sleep
from bs4 import BeautifulSoup
data_folder = '/content/drive/MyDrive/pop_pharmacogenomics/onsides_intl/uk_drug_label/'
external_data_folder = '/content/drive/MyDrive/pop_pharmacogenomics/onsides_intl/external_data/'
api = 'bbf40496-e20e-4f79-9636-b63d4a3deea7'
import warnings
warnings.filterwarnings('ignore')

### extract special population sections and filter for pediatric sections


drug_df = pd.read_csv(data_folder+'data/final_UK/drug_data.csv')
print(drug_df.shape)
drug_df.head(1)

special_pop_text = []
for product in tqdm(drug_df.product_id.unique().tolist()):
  f = (data_folder+'data/raw/{}.txt'.format(product.split('/')[-2]))
  with open(f) as fi:
      s = BeautifulSoup(fi, 'html.parser')
  for i in s.find_all('details'):
      text = i.text.lower()
      if '4.3 contraindications' in text or '4.4 special warnings and precautions for use' in text or '4.6 fertility, pregnancy and lactation' in i.text:
        special_pop_text.append([product.split('/')[-2], text])

special_pop_df = pd.DataFrame(special_pop_text, columns = ['product_id', 'drug_text'])
special_pop_df['section_title'] = special_pop_df.drug_text.apply(lambda x: x.split('\n')[0])
special_pop_df.to_csv(data_folder+'data/drug_special_pop_data_all_raw.csv', index=False)
special_pop_df.head()

special_pop_df = pd.read_csv(data_folder+'data/drug_special_pop_data_all_raw.csv')
special_pop_df['drug_text'] = special_pop_df['drug_text'].apply(lambda x: [i.strip() for i in x.split('\n') if i != ''])
print(special_pop_df.shape)
special_pop_df = special_pop_df.explode('drug_text')
print(special_pop_df.shape)
special_pop_df = special_pop_df[(special_pop_df.drug_text.str.contains('children'))|
                                (special_pop_df.drug_text.str.contains('pediatric'))]
print(special_pop_df.shape)
special_pop_df.to_csv(data_folder+'data/drug_special_pop_data_all_raw_ped_str.csv', index=False)
special_pop_df.head()

### exact string matching

special_pop_df = pd.read_csv(data_folder+'data/drug_special_pop_data_all_raw_ped_str.csv')
print(special_pop_df.shape[0])
unique_terms = special_pop_df[['drug_text']].drop_duplicates()
print(unique_terms.shape[0])
unique_terms.head(1)

f = '/content/drive/MyDrive/pop_pharmacogenomics/onsides_intl/'
meddra_df = pd.read_csv(f+'external_data/umls_meddra_en.csv')
meddra_df['STR'] = meddra_df.STR.apply(lambda x: x.lower())
meddra_df['len'] = meddra_df.STR.apply(lambda x: len(x))
print(meddra_df.shape[0])
meddra_df = meddra_df[(meddra_df['len'] >= 5)]
meddra_df = meddra_df[(meddra_df.TTY.isin(['PT', 'LLT']))]
print(meddra_df.shape[0])
meddra_dict = dict(zip(meddra_df.STR, meddra_df.SDUI))
meddra_df.head(1)

found_ades = []
meddra_names = meddra_df.STR.tolist()
for ade_text in tqdm(unique_terms.drug_text.tolist()):
  ar_text = ' '.join(ade_text.split()).lower()
  found_terms = []
  for concept_name in meddra_names:
    if ar_text.find(concept_name) == -1:
      continue
    else:
      i = ar_text.index(concept_name)
      found_terms.append([i, concept_name, meddra_dict[concept_name]])
  found_ades.append([ade_text, found_terms])

found_ades_df = pd.DataFrame(found_ades, columns = ['string', 'list'])
found_ades_df = found_ades_df.explode('list')
found_ades_df['index'] = found_ades_df['list'].apply(lambda x: x[0] if str(x) != 'nan' else None)
found_ades_df['found_term'] = found_ades_df['list'].apply(lambda x: x[1] if str(x) != 'nan' else None)
found_ades_df['meddra_id'] = found_ades_df['list'].apply(lambda x: x[2] if str(x) != 'nan' else None)
found_ades_df = found_ades_df.drop(['list'], axis = 1).drop_duplicates()
found_ades_df = found_ades_df[found_ades_df['found_term'].notna()]
found_ades_df.to_csv(data_folder+'data/drug_special_pop_data_all_raw_ped_matched.csv', index=False)
found_ades_df.head(1)

---
### format strings for onsides input

found_ades_df = pd.read_csv(data_folder+'data/drug_special_pop_data_all_raw_ped_matched.csv')
special_pop_df = pd.read_csv(data_folder+'data/drug_special_pop_data_all_raw_ped_str.csv')
special_pop_df = special_pop_df.merge(found_ades_df, left_on = 'drug_text', right_on = 'string', how = 'inner')
print(special_pop_df.shape[0], special_pop_df.product_id.nunique(), special_pop_df.found_term.nunique())
special_pop_df.head(1)

building_strings = []
for i, row in tqdm(special_pop_df.iterrows()):
  term, label_id, start_pos = row['found_term'], row['product_id'], int(row['index'])
  #default settings
  nwords, prop_before = 125, 0.125
  #pull the full text
  ar_text = row['drug_text']
  term_nwords = len(term)
  size_before = max(int((nwords-2*term_nwords)*prop_before), 1)
  size_after = max(int((nwords-2*term_nwords)*(1-prop_before)), 1)

  before_text = ar_text[:start_pos]
  after_text = ar_text[(start_pos+term_nwords):]

  before_parts = before_text.split()[-1*size_before:]
  after_parts = after_text.split()[:size_after]

  li = [term]
  li.extend(before_parts)
  li.append('EVENT')
  li.extend(after_parts)
  example_string = ' '.join(li)
  building_strings.append(example_string)
special_pop_df['string'] = building_strings
special_pop_df.head()

#required columns : section, drug, label_id, set_id, spl_version, pt_meddra_id, pt_meddra_term
exact_terms_df = special_pop_df[['product_id','string','found_term','meddra_id']].copy().rename(columns={'product_id':'label_id'})
exact_terms_df['section'] = 'AR'
exact_terms_df['set_id'] = exact_terms_df['label_id']
exact_terms_df['spl_version'] = '0120'

folder = '/content/drive/MyDrive/pop_pharmacogenomics/fda/onsides-2.0.0/data/'
llt_pt = pd.read_csv(folder+'meddra_llt_pt_map.txt', delimiter = '|')
llt_pt_id_dict = dict(zip(llt_pt.llt_concept_code, llt_pt.pt_concept_code))
llt_pt_term_dict = dict(zip(llt_pt.llt_concept_code, llt_pt.pt_concept_name))
exact_terms_df['pt_meddra_id'] = exact_terms_df.meddra_id.apply(lambda x: llt_pt_id_dict[x] if x in llt_pt_id_dict.keys() else None)
exact_terms_df['pt_meddra_term'] =  exact_terms_df.meddra_id.apply(lambda x: llt_pt_term_dict[x] if x in llt_pt_term_dict.keys() else None)

#exact_terms_df.columns = ['section', 'string', 'label_id', 'set_id', 'spl_version', 'pt_meddra_id', 'pt_meddra_term', 'found_term', 'meddra_id']
exact_terms_df.to_csv(data_folder+'data/ped/sentences-rx_method14_nwords125_clinical_bert_application_set_AR_v0120.csv', index=False)
exact_terms_df.head(1)

drug_df = drug_df[['label_id', 'drug']].drop_duplicates().groupby('label_id')['drug'].apply(set).reset_index()
drug_df['drug'] = drug_df['drug'].apply(lambda x: str(x)[1:-1].replace("'",''))
drug_df.head()

exact_terms_df = pd.read_csv(data_folder+'data/ped/sentences-rx_method14_nwords125_clinical_bert_application_set_AR_v0120.csv')
print(exact_terms_df.shape)
drug_df = pd.read_csv(data_folder+'data/final_UK/drug_data.csv').rename(columns={'ingredient_name':'drug'})
drug_df['label_id'] = drug_df['product_id'].apply(lambda x: int(x.split('/')[-2]))
drug_df = drug_df[['label_id', 'drug']].drop_duplicates().groupby('label_id')['drug'].apply(set).reset_index()
drug_df['drug'] = drug_df['drug'].apply(lambda x: str(x)[1:-1].replace("'",''))
exact_terms_df = exact_terms_df.merge(drug_df[['label_id', 'drug']].drop_duplicates(), on = 'label_id', how = 'left')
print(exact_terms_df.shape)
exact_terms_df.to_csv(data_folder+'data/ped/sentences-rx_method14_nwords125_clinical_bert_application_set_AR_v0120.csv', index=False)

exact_terms_df.head()

### run onsides

#prep onsides materials
folder = '/content/drive/MyDrive/pop_pharmacogenomics/onsides_intl/onsides'
ar_model = folder + '/bestepoch-bydrug-PMB_14-AR-125-all_222_24_25_2.5e-05_256_32.pth'
bw_model = folder + '/bestepoch-bydrug-PMB_14-AR-125-all_222_24_25_2.5e-05_256_32.pth'
import os
os.chdir('/content/drive/MyDrive/pop_pharmacogenomics/fda/onsides-2.0.0/')
!python3 -m pip install -r requirements.txt

#use onsides model
f = data_folder+'data/ped/sentences-rx_method14_nwords125_clinical_bert_application_set_AR_v0120.csv'
!python3 src/predict.py --model $ar_model --examples $f

#build tables
r = data_folder+'data/ped/bestepoch-bydrug-PMB-sentences-rx_ref14-AR-125-all_222_24_25_2.5e-05_256_32.csv.gz'
!python3 src/create_onsides_datafiles.py --release v2.0.0-AR --results $r --examples $f
