==> onsides_intl/onsides_uk/notebooks/4.map_drugs_to_rxnorm.ipynb <==
import numpy as np 
import pandas as pd
import requests
from tqdm import tqdm
import ast
from time import sleep
from bs4 import BeautifulSoup
data_folder = ''
api = ''

df = pd.read_csv(data_folder+'drug_data.csv')
df.head(1)

ingredient_df = pd.read_csv(data_folder+'ingredient_data.csv')
print(ingredient_df.shape)
ingredient_df.head()

## Map using Athena RxNorm Data

#filter
athena = '' #external data folder
athena_map = pd.read_csv(athena+'CONCEPT.csv', delimiter = '\t')
athena_map = athena_map[athena_map.vocabulary_id == 'RxNorm']
athena_ingredients = athena_map[athena_map.concept_class_id.isin(['Ingredient', 'Multiple Ingredients', 'Precise Ingredient'])][['concept_id', 'concept_code', 'concept_name']]
athena_ingredients.head()

ingredient_df = ingredient_df.merge(athena_ingredients, left_on = 'ingredient_name', right_on = 'concept_name', how = 'left')
ingredient_df.head()

## UMLS mapping

#use UMLS mapping
for i, row in tqdm(ingredient_df.iterrows()):
  concept_code = row['concept_code'] 
  if str(concept_code) == 'nan':
    ingredient_name = row['ingredient_name']
    url = 'https://uts-ws.nlm.nih.gov/rest/search/current?string={a}&returnIdType=sourceConcept&apiKey={k}'.format(a=ingredient_name, k=api) 
    result = requests.get(url).json()
    if result['result']['recCount'] > 0:
      for r in result['result']['results']:
        if r['rootSource'] == 'RXNORM':
          ingredient_df.at[i, 'concept_code'] = r['ui']
          ingredient_df.at[i, 'concept_name'] = r['name'] 
          print(concept_code, ingredient_name, r['ui'], r['name'])

ingredient_df.to_csv(data_folder+'ingredient_data.csv', index=False)
ingredient_df.head()

## RxNav ID

ingredient_df = pd.read_csv(data_folder+'ingredient_data.csv')
ingredient_df.head()

for i, row in tqdm(ingredient_df.iterrows()):
  if str(row['concept_code']) == 'nan':
    name = row['ingredient_name']
    url = 'https://rxnav.nlm.nih.gov/REST/rxcui.json?name={}&search=0'.format(name.replace(' ', '+'))
    r = requests.get(url).json()
    try:
      rxnorm_id = r['idGroup']['rxnormId'][0]
      ingredient_df.at[i, 'concept_code'] = rxnorm_id
    except:
      pass

for i, row in tqdm(ingredient_df.iterrows()):
  if str(row['concept_code']) == 'nan':
    name = row['ingredient_name']
    url = 'https://rxnav.nlm.nih.gov/REST/rxcui.json?name={}&search=1'.format(name.replace(' ', '+'))
    r = requests.get(url).json()
    try:
      rxnorm_id = r['idGroup']['rxnormId'][0]
      ingredient_df.at[i, 'concept_code'] = rxnorm_id
    except:
      pass

for i, row in tqdm(ingredient_df.iterrows()):
  if str(row['concept_code']) != 'nan' and row['concept_name'] == None:
    code = row['concept_code'] 
    url = 'https://rxnav.nlm.nih.gov/REST/rxcui/{}.json'.format(code)
    r = requests.get(url).json()
    name = r['idGroup']['name']
    ingredient_df.at[i, 'concept_name'] = name

ingredient_df.to_csv(data_folder+'ingredient_data.csv', index=False)
ingredient_df.head()

ingredient_df[ingredient_df.concept_code.isna()]
